+++ History.md
@@ -2,6 +2,13 @@
 
 Please register all relevant modifications to files, JS and state if these modifications took place in a staging, production or local environment, to which branches the files were commited and who did those changes.
 
+December 27th, 2023 | Upload to production | Julieta
+
+-   **Description**: Update components names and fix TUploadFile component. - Upload version 0.0.6
+-   **Main Card**: hthttps://app.clickup.com/t/862kk6p27
+-   **Vue**:
+    -   components/\*
+
 December 21th, 2023 | Upload to production | Julieta
 
 -   **Description**: Disabled option has been added in all form-fields components. - Upload version 0.0.5